# doc/openapi.yaml
openapi: 3.0.3
info:
  title: Tamiyo API
  description: >
    REST API for managing a Magic: The Gathering card collection —
    cards, physical storage (binders, boxes, deckboxes), and decks.
    Every endpoint except /health and /auth/register, /auth/login,
    /auth/refresh, /auth/logout requires a Bearer access token
    (obtained from /auth/register or /auth/login) and scopes all data
    to the authenticated user.
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /cards:
    get:
      tags: [cards]
      summary: List cards
      description: List the authenticated user's cards, with optional filtering and pagination.
      parameters:
        - name: storage_id
          in: query
          required: false
          schema:
            type: integer
          description: Only return cards belonging to this storage.
        - name: name
          in: query
          required: false
          schema:
            type: string
          description: Case-insensitive partial match on the card name.
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          description: 1-based page number.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
          description: Number of cards per page (max 100).
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [name, -name, added, -added, updated, -updated, mana_value, -mana_value]
            default: -updated
          description: >
            Sort field, with an optional "-" prefix for descending order.
            id is always used as a secondary tie-breaker for a stable order.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCards"
        "400":
          description: storage_id, page or limit is invalid, or sort is not one of the allowed values.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [cards]
      summary: Create a card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCardRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        "400":
          description: >
            Missing required field, invalid UUID, invalid collector_number,
            or storage_id does not reference an existing storage owned by
            the authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /cards/{id}:
    parameters:
      - $ref: "#/components/parameters/CardID"
    get:
      tags: [cards]
      summary: Get a card by ID
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        "400":
          $ref: "#/components/responses/InvalidID"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Card not found, or not owned by the authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      tags: [cards]
      summary: Partially update a card
      description: >
        Only the fields present in the body are modified. added and
        updated can never be changed after creation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCardRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        "400":
          description: >
            Invalid id, invalid body, invalid scryfall_id, or storage_id
            does not reference an existing storage owned by the
            authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Card not found, or not owned by the authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [cards]
      summary: Delete a card
      description: >
        Removes the card from any deck it belongs to, and clears
        commander_id on any deck where it was the commander.
      responses:
        "204":
          description: No Content
        "400":
          $ref: "#/components/responses/InvalidID"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Card not found, or not owned by the authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Access token returned by /auth/register, /auth/login or
        /auth/refresh. Short-lived (JWT_ACCESS_TOKEN_TTL_MINUTES,
        15 minutes by default) — use /auth/refresh to obtain a new one
        without logging in again.
  parameters:
    CardID:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: Card ID
...
  responses:
    InvalidID:
      description: id is not a valid integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Unexpected server/database error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing, invalid, or expired access token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Card:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        scryfall_id:
          type: string
          format: uuid
        set_code:
          type: string
        collector_number:
          type: string
        foil:
          type: boolean
        storage_id:
          type: integer
          nullable: true
        mana_value:
          type: number
          description: Converted mana cost (CMC).
        added:
          type: string
        updated:
          type: string
      required:
        - id
        - name
        - scryfall_id
        - set_code
        - collector_number
        - foil
        - mana_value
        - added
        - updated
    PaginatedCards:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Card"
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
      required: [data, page, limit, total, total_pages]
    CreateCardRequest:
      type: object
      properties:
        name:
          type: string
        scryfall_id:
          type: string
          format: uuid
        set_code:
          type: string
        collector_number:
          type: string
        foil:
          type: boolean
          default: false
        storage_id:
          type: integer
          nullable: true
          minimum: 1
        mana_value:
          type: number
          minimum: 0
          default: 0
          description: >
            Converted mana cost (CMC). Not fetched automatically from
            Scryfall — supplied by the client, same as name or set_code.
      required:
        - name
        - scryfall_id
        - set_code
        - collector_number
    UpdateCardRequest:
      type: object
      description: All fields optional; only included fields are modified.
      properties:
        name:
          type: string
        scryfall_id:
          type: string
          format: uuid
        set_code:
          type: string
        collector_number:
          type: string
        foil:
          type: boolean
        storage_id:
          type: integer
          minimum: 1
        mana_value:
          type: number
          minimum: 0
...